requirements.txt
eth_account==0.11.2
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.30
Werkzeug==3.0.3
stripe==10.12.0
Flask-Mail==0.10.0
base58==2.1.1
pynacl==1.5.0